+++ HackerRank/Algorithms/Implementation/InsertionSortPart1.hpp
@@ -24,7 +24,7 @@ namespace InsertionSortPart1 {
         
         int key = arr.back();
         
-        for(int i = arr.size() - 1; i >= 0; --i) {
+        for(int i = (int)arr.size() - 1; i >= 0; --i) {
             if(key < arr[i - 1])
                 arr[i] = arr[i-1];
             else{
